requirements.txt
Flask==2.2.2
Flask-JWT-Extended==4.4.4
Flask-SQLAlchemy==3.0.2
psycopg2-binary
pulsar-client
requests==2.28.1
python-dotenv==0.21.0
Werkzeug==2.2.2
flask-swagger==0.2.13
Flask-Limiter
redis
flask-redis
PyDispatcher
dispatcher  
pydispatch
psycopg2
pulsar-client[avro]